- Renaming UserProfile into PlayerAuthentication - Moved all scripts into the Xbox Live namespace - Fixed several bugs in authentication - Applied a few changes per PR feedback
