+++ Assets/Xbox Live/Examples/Social.unity
@@ -38,7 +38,7 @@ RenderSettings:
   m_ReflectionIntensity: 1
   m_CustomReflection: {fileID: 0}
   m_Sun: {fileID: 0}
-  m_IndirectSpecularColor: {r: 0.44657874, g: 0.49641258, b: 0.5748172, a: 1}
+  m_IndirectSpecularColor: {r: 0.44657844, g: 0.49641222, b: 0.57481694, a: 1}
 --- !u!157 &3
 LightmapSettings:
   m_ObjectHideFlags: 0
@@ -247,22 +247,22 @@ Prefab:
     - target: {fileID: 224363205643076876, guid: 35fb2ee2a61d5ca4eafc9ee1e66325e0,
         type: 2}
       propertyPath: m_AnchoredPosition.x
-      value: -242
+      value: -220
       objectReference: {fileID: 0}
     - target: {fileID: 224363205643076876, guid: 35fb2ee2a61d5ca4eafc9ee1e66325e0,
         type: 2}
       propertyPath: m_AnchoredPosition.y
-      value: 0
+      value: -58
       objectReference: {fileID: 0}
     - target: {fileID: 224363205643076876, guid: 35fb2ee2a61d5ca4eafc9ee1e66325e0,
         type: 2}
       propertyPath: m_SizeDelta.x
-      value: 600
+      value: 458
       objectReference: {fileID: 0}
     - target: {fileID: 224363205643076876, guid: 35fb2ee2a61d5ca4eafc9ee1e66325e0,
         type: 2}
       propertyPath: m_SizeDelta.y
-      value: 400
+      value: 425
       objectReference: {fileID: 0}
     - target: {fileID: 224363205643076876, guid: 35fb2ee2a61d5ca4eafc9ee1e66325e0,
         type: 2}
@@ -436,114 +436,114 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 238}
   m_SizeDelta: {x: 216, y: 40}
   m_Pivot: {x: 0.5, y: 0.5}
---- !u!1001 &1105474959
+--- !u!1001 &1303126072
 Prefab:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Modification:
     m_TransformParent: {fileID: 0}
     m_Modifications:
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_LocalPosition.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_LocalPosition.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_LocalPosition.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_LocalRotation.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_LocalRotation.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_LocalRotation.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_LocalRotation.w
       value: 1
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_RootOrder
-      value: 1
+      value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_AnchoredPosition.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_AnchoredPosition.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_SizeDelta.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_SizeDelta.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_AnchorMin.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_AnchorMin.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_AnchorMax.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_AnchorMax.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_Pivot.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_Pivot.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_LocalScale.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_LocalScale.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 224148228764223538, guid: f9c76770c5e766a4f83203b145924f14,
+    - target: {fileID: 224507845527857528, guid: f9c76770c5e766a4f83203b145924f14,
         type: 2}
       propertyPath: m_LocalScale.z
       value: 0
@@ -697,6 +697,11 @@ Prefab:
     - target: {fileID: 4896899429925906, guid: 2b8f49ee71ade254da6555182196f225, type: 2}
       propertyPath: m_RootOrder
       value: 2
+      objectReference: {fileID: 0}
+    - target: {fileID: 114670622419244234, guid: 2b8f49ee71ade254da6555182196f225,
+        type: 2}
+      propertyPath: DebugLogsOn
+      value: 0
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 100100000, guid: 2b8f49ee71ade254da6555182196f225, type: 2}
