+++ Assets/Xbox Live/Scripts/AvailabilityCheck.cs
@@ -3,19 +3,23 @@
 // 
 using UnityEngine;
 
-public class AvailabilityCheck : MonoBehaviour {
-    // The UI element to display when the device is offline
-    public GameObject OfflineStatus;
-
-    bool showingOffline;
-
-    void Update()
+namespace Microsoft.Xbox.Services.Client
+{
+    public class AvailabilityCheck : MonoBehaviour
     {
-        bool offline = Application.internetReachability == NetworkReachability.NotReachable;
-        if (showingOffline != offline)
+        // The UI element to display when the device is offline
+        public GameObject OfflineStatus;
+
+        bool showingOffline;
+
+        void Update()
         {
-            OfflineStatus.SetActive(offline);
-            showingOffline = offline;
+            bool offline = Application.internetReachability == NetworkReachability.NotReachable;
+            if (showingOffline != offline)
+            {
+                OfflineStatus.SetActive(offline);
+                showingOffline = offline;
+            }
         }
     }
 }
