+++ Assets/Xbox Live/Scripts/LeaderboardEntry.cs
@@ -7,72 +7,75 @@ using System.Collections;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class LeaderboardEntry : MonoBehaviour
+namespace Microsoft.Xbox.Services.Client
 {
-    public Button entryButton;
-    public Image gamerpicImage;
-    public Text rankText;
-    public Text gamertagText;
-    public Text valueText;
-
-    private LeaderboardRow data;
-
-    public LeaderboardRow Data
+    public class LeaderboardEntry : MonoBehaviour
     {
-        get
-        {
-            return this.data;
-        }
-        set
-        {
-            this.data = value;
-            this.gamertagText.text = this.data.Gamertag;
-            this.rankText.text = this.data.Rank.ToString();
-            this.valueText.text = this.data.Values[0];
-            this.gamerpicImage.sprite = null;
-        }
-    }
+        public Button entryButton;
+        public Image gamerpicImage;
+        public Text rankText;
+        public Text gamertagText;
+        public Text valueText;
 
-    string gamerpicUrl;
-    public string GamerpicUrl
-    {
-        get
-        {
-            return this.gamerpicUrl;
-        }
-        set
-        {
-            this.gamerpicUrl = value;
-            StartCoroutine(LoadGamerpic());
-        }
-    }
-    IEnumerator LoadGamerpic()
-    {
-        var www = new WWW(gamerpicUrl);
-        yield return null;
+        private LeaderboardRow data;
 
-        while (!www.isDone)
+        public LeaderboardRow Data
         {
-            yield return new WaitForSeconds(0.2f);
-        }
-
-        try
-        {
-            if (!string.IsNullOrEmpty(www.error))
+            get
             {
-                throw new Exception(www.error);
+                return this.data;
             }
-                
-            var t = www.texture;
-            var r = new Rect(0, 0, t.width, t.height);
-            this.gamerpicImage.color = Color.white;
-            this.gamerpicImage.sprite = Sprite.Create(t, r, Vector2.zero);
-        }
-        catch (Exception ex)
-        {
-            if (XboxLiveServicesSettings.Instance.DebugLogsOn)
+            set
             {
-                Debug.Log("There was an error while loading " + gamertagText.text + "'s gamerpic. Exception: " + ex.Message);
+                this.data = value;
+                this.gamertagText.text = this.data.Gamertag;
+                this.rankText.text = this.data.Rank.ToString();
+                this.valueText.text = this.data.Values[0];
+                this.gamerpicImage.sprite = null;
+            }
+        }
+
+        string gamerpicUrl;
+        public string GamerpicUrl
+        {
+            get
+            {
+                return this.gamerpicUrl;
+            }
+            set
+            {
+                this.gamerpicUrl = value;
+                StartCoroutine(LoadGamerpic());
+            }
+        }
+        IEnumerator LoadGamerpic()
+        {
+            var www = new WWW(gamerpicUrl);
+            yield return null;
+
+            while (!www.isDone)
+            {
+                yield return new WaitForSeconds(0.2f);
+            }
+
+            try
+            {
+                if (!string.IsNullOrEmpty(www.error))
+                {
+                    throw new Exception(www.error);
+                }
+
+                var t = www.texture;
+                var r = new Rect(0, 0, t.width, t.height);
+                this.gamerpicImage.color = Color.white;
+                this.gamerpicImage.sprite = Sprite.Create(t, r, Vector2.zero);
+            }
+            catch (Exception ex)
+            {
+                if (XboxLiveServicesSettings.Instance.DebugLogsOn)
+                {
+                    Debug.Log("There was an error while loading " + gamertagText.text + "'s gamerpic. Exception: " + ex.Message);
+                }
             }
         }
     }
