+++ Assets/Xbox Live/Scripts/LeaderboardTypes.cs
@@ -1,7 +1,12 @@
 ﻿// Copyright (c) Microsoft Corporation
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-public enum LeaderboardTypes{
-    Global,
-    Friends,
-    Favorites
+
+namespace Microsoft.Xbox.Services.Client
+{
+    public enum LeaderboardTypes
+    {
+        Global,
+        Friends,
+        Favorites
+    }
 }
