+++ Assets/Xbox Live/Scripts/Stats/StatEventArgs.cs
@@ -2,11 +2,9 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 // 
 using System;
-
-using Microsoft.Xbox.Services;
 using Microsoft.Xbox.Services.Statistics.Manager;
 
-namespace XboxLivePrefab
+namespace Microsoft.Xbox.Services.Client
 {
     public class StatEventArgs : EventArgs
     {
