+++ Assets/Xbox Live/Scripts/XboxLiveServicesSettings.cs
@@ -5,26 +5,34 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class XboxLiveServicesSettings : Singleton<XboxLiveServicesSettings> {
-
-    public bool DebugLogsOn = true;
-
-    private void Awake()
+namespace Microsoft.Xbox.Services.Client
+{
+    public class XboxLiveServicesSettings : Singleton<XboxLiveServicesSettings>
     {
-        // Ensure that a XboxLiveServicesSettings Instance has been created.
-        var manager = XboxLiveServicesSettings.Instance;
-        DontDestroyOnLoad(this);
-    }
 
+        public bool DebugLogsOn = true;
 
-    /// <summary>
-    /// Ensures that there is an Xbox Live Debug Manager instance on the scene used to configure debug settings of Xbox Live Prefabs
-    /// </summary>
-    public static void EnsureXboxLiveServicesSettings()
-    {
-        if (Object.FindObjectOfType<XboxLiveServicesSettings>() == null)
+        private void Awake()
         {
-            Debug.LogErrorFormat("Make sure to drag at least one instance of the XboxLiveServices prefab on your initial scene.");
+            // Ensure that a XboxLiveServicesSettings Instance has been created.
+            var manager = XboxLiveServicesSettings.Instance;
+            DontDestroyOnLoad(this);
+        }
+
+
+        /// <summary>
+        /// Ensures that there is an Xbox Live Debug Manager instance on the scene used to configure debug settings of Xbox Live Prefabs
+        /// </summary>
+        public static void EnsureXboxLiveServicesSettings()
+        {
+            if (Object.FindObjectOfType<XboxLiveServicesSettings>() == null)
+            {
+                Debug.LogErrorFormat("Make sure to drag at least one instance of the XboxLiveServices prefab on your initial scene.");
+            }
+        }
+
+        public static bool EnsureXboxLiveServiceConfiguration() {
+            return XboxLive.Instance.AppConfig != null && XboxLive.Instance.AppConfig.ServiceConfigurationId != null;
         }
     }
 }
