--- Assets/Xbox Live/Scripts/XboxLiveUserInfo.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: d55824f79eac22642a2cba22c1e4da5e
-timeCreated: 1496430280
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
