--- Assets/Xbox Live/Scripts/XboxLiveUserManager.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: c4fc7c07612376745a80de65062e9bf4
-timeCreated: 1496430166
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
